requirements.txt
Flask~=3.0.3
langchain~=0.3.3
openai
chromadb
pypdf
python-dotenv
langchain-community
tiktoken
langchain-openai
Flask-cors
gunicorn
pillow
Werkzeug